feat(测试跟踪): 测试计划增加批量删除操作
--story=1010195 --user=李玉号 【测试计划】测试计划新增批量删除功能
https://www.tapd.cn/55049933/s/1291099

+++ test-track/backend/src/main/java/io/metersphere/plan/service/TestPlanPrincipalService.java
@@ -3,11 +3,13 @@ package io.metersphere.plan.service;
 import io.metersphere.base.domain.TestPlanPrincipal;
 import io.metersphere.base.domain.TestPlanPrincipalExample;
 import io.metersphere.base.mapper.TestPlanPrincipalMapper;
+import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
+import java.util.List;
 
 @Service
 @Transactional(rollbackFor = Exception.class)
@@ -23,6 +25,15 @@ public class TestPlanPrincipalService {
         }
         TestPlanPrincipalExample example = new TestPlanPrincipalExample();
         example.createCriteria().andTestPlanIdEqualTo(planId);
+        testPlanPrincipalMapper.deleteByExample(example);
+    }
+
+    public void deletePlanPrincipalByPlanIds(List<String> planIds) {
+        if (CollectionUtils.isEmpty(planIds)) {
+            return;
+        }
+        TestPlanPrincipalExample example = new TestPlanPrincipalExample();
+        example.createCriteria().andTestPlanIdIn(planIds);
         testPlanPrincipalMapper.deleteByExample(example);
     }
